Extract cookie control logic to route handler
it is a good practice to separate business logic from rendering logic. Here, I extracted logic for setting cookie and deleting cookie into the handlers for sign in and sign out, so that the header component and signin page UI does not know about where cookie is being set. Moreover, it reduce redundant use case for using cookie library